backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pytest==7.4.3
httpx==0.25.2
python-multipart==0.0.6